requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
sqlalchemy==2.0.30
aiosqlite==0.20.0
psycopg[binary]==3.1.18
argon2-cffi==23.1.0
PyJWT==2.8.0
email-validator==2.1.1
python-multipart==0.0.9
